+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * faers_meta(internal = TRUE) will always use the cache data in the package. 
 
+* Rename "gndr_cod" into "gender" for periods before 2014q2, and rename "sex" into "gender" for periods after or equal to 2014q2. 
+
+* "sex" was added, which recoded any values other than "F" or "M" as `NA`.
+
 # faers 1.1.4
 
 * fix error when download failed